+++ R/datasets.R
@@ -239,8 +239,8 @@
 #' emmip(nutr.emm, race ~ group)
 #' 
 #' # Hispanics seem exceptional; but this doesn't test out due to very sparse data
-#' CLD(nutr.emm, by = "group")
-#' CLD(nutr.emm, by = "race")
+#' pairs(nutr.emm, by = "group")
+#' pairs(nutr.emm, by = "race")
 "nutrition"