+++ R/helpers.R
@@ -38,7 +38,9 @@ emm_basis.lm = function(object, trms, xlev, grid, ...) {
     bhat = object$coefficients
     nm = if(is.null(names(bhat))) row.names(bhat) else names(bhat)
     m = suppressWarnings(model.frame(trms, grid, na.action = na.pass, xlev = xlev))
-    X = model.matrix(trms, m, contrasts.arg = object$contrasts)[, nm, drop = FALSE]
+    X = model.matrix(trms, m, contrasts.arg = object$contrasts)
+    assign = attr(X, "assign")
+    X = X[, nm, drop = FALSE]
     bhat = as.numeric(bhat) 
     # stretches it out if multivariate - see mlm method
     V = .my.vcov(object, ...)
@@ -57,7 +59,8 @@ emm_basis.lm = function(object, trms, xlev, grid, ...) {
         dfargs = list(df = object$df.residual)
         dffun = function(k, dfargs) dfargs$df
     }
-    list(X=X, bhat=bhat, nbasis=nbasis, V=V, dffun=dffun, dfargs=dfargs, misc=misc)
+    list(X=X, bhat=bhat, nbasis=nbasis, V=V, dffun=dffun, dfargs=dfargs, misc=misc,
+         model.matrix = .cmpMM(object$qr, assign = assign))
 }