+++ R/ref-grid.R
@@ -690,6 +690,8 @@ ref_grid <- function(object, at, cov.reduce = mean, cov.keep = get_emm_option("c
     grid[[".wgt."]] = wgt
     
     model.info = list(call = attr(data,"call"), terms = trms, xlev = xlev)
+    model.info$model.matrix = basis$model.matrix  # for submodel support
+    model.info$ylevs = basis$ylevs
     # Detect any nesting structures
     nst = .find_nests(grid, trms, coerced$orig, ref.levels)
     if (length(nst) > 0)