+++ R/zzz.R
@@ -25,8 +25,9 @@
 # maybe we can we require R >= 3.4 (first that has hasName())
 # and add utils::hasName to imports (in emmeans-package.R)
 
-hasName = function(x, name)
-   match(name, names(x), nomatch = 0L) > 0L
+### No longer needed as now I require R >= 3.5.0
+# hasName = function(x, name)
+#    match(name, names(x), nomatch = 0L) > 0L