+++ NEWS.md
@@ -9,6 +9,11 @@ emmeans 1.4.809xxxxxxx
   * In `emmeans()` with a `contr` argument or two-sided formula, we now suppress
     several particular `...` arguments from being passed on to `contrast()`
     when they should apply only to the construction of the EMMs (#214)
+  * More control of what `...` arguments are passed to methods
+  * 'CLD()` was deprecated in version 1.3.4, now it's gone. 
+    Documentation for `multcomp::cld()`
+    method added and moved to different help page.
+  * Experimental `submodel` option